simplify scrolling setup
pixel-scroll-precision-mode doesn't work with scroll-margin currently. it's nice to scroll with
cursor in the center, but that can still be done with `scroll-up-command` and `scroll-down-command`
and the occassional `recenter-top-bottom`. it is annoying when `scroll-margin` kicks in for
compilation output sorts of buffers too, so just trying to simplify for now